Добавил сервер который принимает get-запрос и отдает ответ
