+++ hw15_go_sql/main.go
@@ -2,15 +2,23 @@ package main
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"log"
+	"net/http"
 	"os"
 
 	"github.com/jackc/pgx/v5/pgxpool"
+	"github.com/spf13/pflag"
 )
 
-func old() {
-	ctx := context.Background()
+type Person struct {
+	Name    string
+	Surname string
+}
+
+func getDatabaseConn() (ctx context.Context, conn *pgxpool.Pool) {
+	ctx = context.Background()
 	dsn := "postgres://postgres:postgres@localhost:5432/test_db?search_path=test_schema&sslmode=disable&pool_max_conns=20"
 
 	pgCfg, err := pgxpool.ParseConfig(dsn)
@@ -19,130 +27,62 @@ func old() {
 	}
 
 	// conn, err := pgxpool.New(ctx, dsn)
-	conn, err := pgxpool.NewWithConfig(ctx, pgCfg)
+	conn, err = pgxpool.NewWithConfig(ctx, pgCfg)
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "Unable to connect to database: %v\n", err)
 		os.Exit(1)
 	}
 
-	defer conn.Close()
-
 	if err = conn.Ping(ctx); err != nil {
 		log.Fatal("We cannot connect to database")
 	}
+	return
+}
 
-	// Вывод всех пользователей
-	users, err := GetUsers(ctx, conn)
-	if err != nil {
-		log.Fatal(err)
-	}
-	for _, user := range users {
-		fmt.Printf("ID: %d, Имя: %s, Email: %s, Пароль: %s\n", user.ID, user.Name, user.Email, user.Password)
-	}
+func getParams() (address string, port int16) {
+	// Когда указывают с именами: go run main.go -u=localhost -p=8080
+	pflag.StringVarP(&address, "address", "u", "localhost", "server address")
+	pflag.Int16VarP(&port, "port", "p", 8080, "service port")
 
-	// Вывод всех заказов
-	orders, err := GetUserOrders(ctx, conn)
-	if err != nil {
-		log.Fatal(err)
-	}
-	for _, order := range orders {
-		fmt.Printf(
-			"ID: %d, ID пользователя: %d, Дата заказа: %s, Сумма: %f\n",
-			order.ID, order.UserID, order.OrderDate, order.TotalAmount,
-		)
+	pflag.Parse()
+
+	// Когда указывают: go run main.go - подставляем переменные окружения
+	// -o будет равен значению из переменной окружения если его вообще никак не укажут
+	if address == "" && port == 0 {
+		log.Fatal("Ошибка")
 	}
 
-	// Вывод статистики по пользователю
-	userStats, err := GetUserStat(ctx, conn)
-	if err != nil {
-		log.Fatal(err)
-	}
-	for _, userStat := range userStats {
-		fmt.Printf(
-			"Имя: %s, Сумма заказов: %f, Средняя цена товара: %f\n",
-			userStat.UserName, userStat.TotalAmount, userStat.AvgPrice,
-		)
-	}
+	return
+}
 
-	// Добавление пользователя
-	name := "Vladislav"
-	email := "Vladislav@mail.ru"
-	password := "Vladislav1234"
-	err = AddUser(ctx, conn, name, email, password)
-	if err != nil {
-		log.Fatal(err)
+func getProducts(w http.ResponseWriter, r *http.Request) {
+	switch r.Method {
+	case "GET":
+		// Вывод всех продуктов
+		ctx, conn := getDatabaseConn()
+		defer conn.Close()
+		products, err := GetProducts(ctx, conn)
+		if err != nil {
+			log.Fatal(err)
+		}
+		b, err := json.Marshal(products)
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
+		result := string(b)
+		log.Println(result)
+		fmt.Fprintf(w, result)
+	default:
+		code := 405
+		http.Error(w, "Only GET method supported.", code)
 	}
-	fmt.Println("Пользователь добавлен успешно.")
+}
 
-	// Обновление пользователя
-	id := 4
-	name = "Борис"
-	email = "boris@mail.ru"
-	password = "dasf"
-	err = UpdateUser(ctx, conn, id, name, email, password)
-	if err != nil {
-		log.Fatal(err)
-	}
-	fmt.Println("Пользователь обновлен успешно.")
+func main() {
+	address, port := getParams()
+	host := fmt.Sprintf("%s:%d", address, port)
 
-	// Удаление пользователя
-	id = 6
-	err = DeleteUser(ctx, conn, id)
-	if err != nil {
-		log.Fatal(err)
-	}
-	fmt.Println("Пользователь удален успешно.")
-
-	// Добавление продукта
-	name = "Сок"
-	price := float32(90.0)
-	err = AddProduct(ctx, conn, name, price)
-	if err != nil {
-		log.Fatal(err)
-	}
-	fmt.Println("Продукт добавлен успешно.")
-
-	// Обновление продукта
-	id = 2
-	name = "Йогурт"
-	price = 64.0
-	err = UpdateProduct(ctx, conn, id, name, price)
-	if err != nil {
-		log.Fatal(err)
-	}
-	fmt.Println("Продукт обновлен успешно.")
-
-	// Удаление продукта
-	id = 5
-	err = DeleteProduct(ctx, conn, id)
-	if err != nil {
-		log.Fatal(err)
-	}
-	fmt.Println("Продукт удален успешно.")
-
-	// Добавление заказа
-	userID := 3
-	totalAmount := float32(500.0)
-	err = AddOrder(ctx, conn, userID, totalAmount)
-	if err != nil {
-		log.Fatal(err)
-	}
-	fmt.Println("Заказ добавлен успешно.")
-
-	// Добавление продуктов заказа
-	orderID := 2
-	productID := 2
-	err = AddOrderProduct(ctx, conn, orderID, productID)
-	if err != nil {
-		log.Fatal(err)
-	}
-	fmt.Println("Продукты заказа добавлены успешно.")
-
-	// Удаление заказа
-	id = 1
-	err = DeleteOrder(ctx, conn, id)
-	if err != nil {
-		log.Fatal(err)
-	}
-	fmt.Println("Заказ удален успешно.")
+	http.HandleFunc("/get_products", getProducts)
+	http.ListenAndServe(host, nil) //nolint
 }
